fix: Create hashtag filters with the `#` character
Previous code created hashtag filters without the `#`, so muting the
tag `#something` would filter all posts that contained `something`,
with or without the `#`.

Fix this when creating filters, and only remove filters (when unmuting)
if the title and contents match.

Show a snackbar when hashtags are successfully muted/unmuted, so the
user is aware something happened.