docs: Expand translate/code guides for new contributors
Instead of a single CONTRIBUTING.md, create an initial set of more
detailed documentation for the different types of contribution.

Start with `code.md` and `translate.md`, the processes for submitting
code and translations respectively.

Commit to documenting how to contribute documentation, project
management, user support, and social presence in the next week.