Comment out the reading position restoration code
In the most recent tests the reading position after re-opening the app
appeared to briefly be correct. It was then set to a position 30
posts after the saved reading position.

So maybe the manual restoration is getting in the way of work the
adapater/recyclerview is doing to restore the position?