Try with a larger jump threshold
Set jumpThreshold to 3x the page size. That should allow it to jump
if the initial loaded page is larger.

Put back deferredPrevPage, assuming that jumpThreshold will now find
the status-to-restore at position 31.

Put back the shouldRestoreReadingPosition code, so the reading position
is restored if the app is restarted from scratch.