fix: Restore the user's reading position under all circumstances
The previous code did not always work when the user returned to the
app after a lengthy absence (e.g., overnight).

Instead of restoring by scrolling in `TimelineFragment`,
restore by working with the platform.

Determine the initial page to fetch by looking half a page ahead of the
saved saved status ID, and fetch that status and the page immediately
prior. This seems to match the view's expectations about what will be
immediately available.

Set `jumpThreshold` and `enablePlaceholders` in the `PagingConfig` so
the paging system will jump to the saved status.

Remove the restoration code in `TimelineFragment`.

Fixes #53