feat: Embed the privacy policy in the app
Instead of linking to the privacy policy embed it in the app as a
string of HTML.

The string is created with a new `markdown2resource` plugin, which
converts `PRIVACY.md` to HTML and generates a Java class with the HTML
content.

Create `PrivacyPolicyActivity` to display the HTML in a `WebView`,
and link to it from `AboutActivity`.