feat: show announcement dates(#35)
Previously, it was inconvenient to know when "Announcements" were announced. So, display the date when "Announcements" was published on the screen(This should also correspond to relative and absolute time).

Also, when "Announcements" is updated, it is inconvenient if the date of the update is not displayed. Therefore, display the date when "Announcements" is updated on the screen.