fix: Quit using `toString()` for date comparisons between `updatedAt` and `publishedAt` (#151)
The previous code used `toString()` for date comparison between publishAt and updatedAt of "Announcements".
Remove `toString()` for simple date comparisons in the Date class since there is no need to use it

However, when stopping the use of `toString()`, the "updated" indicator appeared on the screen even though updatedAt and publishedAt appeared to be at the same time. This was because there was a time lag of less than one minute.
Therefore, Create an extension function `equalByMinute` of Date class for date comparison that ignores the time lag between updatedAt and publishedAt that is less than one minute.

Fixes #151

----------
Co-authored-by: Nik Clayton <nik@ngo.org.uk>