View binding for everything except StatusBaseViewHolder
Add FilterableStatusViewHolder, move the filtering logic there (out of
StatusBaseViewHolder).

Because view binding classes can't inherit from one another this means
that StatusViewHolder also has to use ItemStatusWrapperBinding. This
adds an additional level to each status view in a list which might
not be good, so this is an experiment.