fix: Use string resource to display "Announcements" dates (#151)
Previously, the "Announcements" date was displayed by hard-coding a string.
However, using string resource to display text allows for more flexibility, such as the ability to change the order in which text appears depending on the translation.

Therefore, use string resource for displaying "Announcements" date as well as `reaction_name_and_count`.

Fixes #151