refactor: Use sealed interfaces for lighterweight code
Using a sealed interface (instead of a sealed class) at the root of
the hierarchy avoids the overhead of having to create and
initialise the class (visible in the generated bytecode).

It also makes the instantiation code slightly less cumbersome because
the code doesn't need to pass parameters to the root's constructor.