refactor: Change the name of the preference key (#193)
The previous code was incorrect in naming the preference key.
If the value of preference is a boolean value, it is reasonable to assume that the filter is applied if the value is true because the name contains "filter".

Therefore, change "filter" to "show" in the preference key name.

Fixes #193

----------
Co-authored-by: Nik Clayton <nik@ngo.org.uk>