change: Turn off the "Show self-boosts" switch when the "Show boosts" switch is off (#193)
The previous code allowed the "Show self-boosts" switch to be toggled even when the "Show boosts" switch was off.
However, when the "Show boosts" switch is off, "Show self-boosts" must always be off.

Therefore, change it so that if the user disables all boosts, the "Disable self-boosts" setting is ignored.

Fixes #193

----------
Co-authored-by: Nik Clayton <nik@ngo.org.uk>