fix: Remove progressbar from status timelines
Previously the middle-of-screen progress spinner and the spinner that
appears on a swipe-to-refresh could get out of sync.

Fix this by removing the middle-of-screen progress spinner from
relevant fragments, as the swipe-to-refresh spinner shows the user
that an operation is in progress, and also clues them in to the fact
that a swipe-to-refresh is possible (by using the common UX control).

Fixes #75