fix: Ensure system controls are visible on older devices
Previous code always set `navigationBarColor` and `statusBarColor` to
`transparent` irrespective of the API level.

This only works on API 29 and above; if you do it on API levels lower
than that the system navigation buttons (home, back, recents) are
typically shown on a very similar colour to the background, making them
very hard to see.

Fixes #221