feat: Periodically check for updates and alert user
Users can inadvertently get stuck on older versions of the app; e.g.,
by installing from one F-Droid repository that stops hosting the app at
some later time.

Analytics from the Play Store also shows a long tail of users who are,
for some reason, on an older version.

On resuming `MainActivity`, and approximately once per day, check and
see if a newer version of Pachli is available, and prompt the user to
update by going to the relevant install location (Google Play, F-Droid,
or GitHub).

The dialog prompt allows them to ignore this specific version, or
disable all future update notifications. This is also exposed through
the preferences, so the user can adjust it there too.

A different update check method is used for each installation location.

- F-Droid: Use the F-Droid API to query for the newest released version
- GitHub: Use the GitHub API to query for the newest release, and
  check the APK filename attached to that release
- Google Play: Use the Play in-app-updates library
  (https://developer.android.com/guide/playcore/in-app-updates) to
  query for the newest released version

These are kept in different build flavours (source sets), so that
e.g., the build for the F-Droid store can only query the F-Droid API,
the UI strings are specific to F-Droid, etc. This also ensures that
the update service libraries are specific to that build and do not
"cross-contaminate".

Note that this *does not* update the app, it takes the user to
either the relevant store page (F-Droid, Play) or GitHub release
page. The user must still start the update from that page.

CI configuration is updated to build the different flavours.