feat: Set the initial cursor position of EditText to the left edge when the compose button is clicked in a hashtag timeline
The previous code positioned the cursor in EditText at the right edge when the compose button was pressed from the hashtag timeline. So This was inconvenient because the cursor had to be moved to the left edge in order to enter text.

Therefore, solve this by positioning the initial position of the cursor at the left end.

Fixes #254, #228

----------------------
Co-authored-by: Nik Clayton <nik@ngo.org.uk>