fix: Set the cursor position to the right edge of EditText when creating a reply.
The previous code positioned the EditText cursor to the left when creating a reply.
However, the cursor in Edittext should only be positioned on the left edge when creating a toot in the hashtag timeline.

Therefore, set the cursor of EditText to be positioned at the right edge when creating a reply.

Fixes #254, #228

----------------------
Co-authored-by: Nik Clayton <nik@ngo.org.uk>