fix: Show correct trending tag values at the end of the chart lines
The previous code incorrectly showed the trending tag usage data twice
next to the end of the trending tag lines, instead of one entry for
the usage data and one entry for the account data.

Fix that.

As part of this fix change how the data is displayed. Instead of using
two distinct `TextView`, fixed to the bottom end of the chart, draw
the text directly on the chart. The text is accurately position so that
it is next to the end of the relevant line. If both lines overlap the
label positions are adjusted appropriately.

While doing this I discovered that the mechanism used to fall back to
particular chart colours if none were specified was incorrect, so fix
that too.