fix: Allow users to retry errors when fetching server info
Previous code would show errors when fetching server info but with
no mechanism to retry the operation, so it would "stick" until the user
restarted the app.

Fix this with a retry action.

While I'm here use styles to ensure that all snackbars are 5 lines max
length and remove code that sets the length.