change: Display "About" information in tabs
Previously, `AboutActivity` had buttons and links to show the privacy
policy and licenses of dependencies.

Change this to a selection of fragments in tabs, one tab each for:

- General "About" information
- Licenses
- Privacy Policy

The information shown hasn't changed, but this lays the groundwork
for including additional tabs in the future for information like
server rules, detected capabilities, or troubleshooting information.