feat: Show information about notification fetches on "About" screen
Some users report that Pachli is not retrieving/displaying notifications
in a timely fashion.

To assist in diagnosing these errors, provide an additional set of tabs
on the "About" screen that contain information about how Pachli is
fetching notifications, and if not, why not.

Allow the user to save notification related logs and other details to
a file that can be attached to an e-mail or bug report.

Recording data:

- Provide a `NotificationConfig` singleton with properties to record
  different aspects of the notification configuration. Update these
  properties as different notification actions occur.

- Store logs in a `LogEntryEntity` table. Log events of interest with
  a new `Timber` `LogEntryTree` that is planted in all cases.

- Add `PruneLogEntryEntityWorker` to trim saved logs to the last 48
  hours.

Display:

- Add a `NotificationFragment` to `AboutActivity`. It hosts two other
  fragments in tabs to show details from `NotificationConfig` and the
  relevant logs, as well as controls for interacting with them.

Bug fixes:

- Filter out notifications with a null tag when processing active
  notifications, prevents an NPE crash

Other changes:

- Log more details when errors occur so the bug reports are more helpful