fix: Prevent crash when Friendica returns a null `voted_on` property
Friendica can return a null `voted_on` property, in violation of the
API spec.

Introduce a `BooleanIfNull` annotation that will convert the `null`
to `false` if encountered.

While I'm here update the other adapters as classes on their
relevant annotations instead of standalone classes to keep the code
consistent.

Fixes #455