ci: Add workflow to build APK from PR
Debugging sometimes requires shipping APKs that have additional
instrumentation to users.

It's bad practice to encourage users to install APKs that they have
been given by a stranger on the Internet.

This new workflow will build a properly signed APK from a PR when a
specific comment is added to the PR. After building a comment is
added with a link to the workflow run and the URL the user can use
to download the APK. This allows them to be sure that the APK contains
only the code in the PR.