feat: Include "Notifications" in the left-side navigation menu
Previously the only way to access notifications was to dedicate a tab
to them. Now notifications are available from the left-side navigation
menu so they're always accessible.

Add them to the top of the list, and swap the order of bookmarks and
favourites, assuming that users are more likely to want to see their
bookmarks than their favourites.

Move "Edit profile" to the bottom with the other settings options,
assuming that editing their profile does not happen very often, so
should not be at the top of the list.