fix: Ensure logging out accounts completes
The account logout process could fail due to API exceptions; network
errors for example, or if the user had already revoked the app's
token for that account. This would prevent the rest of the logout
process (cleaning database, etc) from completing.

Fix this by ignoring network errors during the logout process, and
always cleaning up account content in the database.

Fix a related issue where a deleted account might be recreated in a
partial state if the account's visible position was saved after it
was deleted. The recreated account couldn't do anything as it had no
tokens, but is very confusing.